pyproject.toml
[project]
name = "Personal-Finance-Tracker"
version = "0.1.0"
description = "Personal-Finance-Tracker"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.116.0",
    "httpx>=0.28.1",
    "langchain-groq>=0.3.5",
    "langchain-openai>=0.3.27",
    "load-dotenv>=0.1.0",
    "mcp-use>=1.3.6",
    "uvicorn>=0.35.0",
]
